+++ aten/src/ATen/native/AdaptivePooling.h
@@ -2,6 +2,7 @@
 
 #include <ATen/core/Tensor.h>
 #include <ATen/native/DispatchStub.h>
+#include <ATen/native/PoolingChecks.h>
 #include <c10/util/ArrayRef.h>
 #include <c10/util/irange.h>
 #include <cmath>
@@ -34,16 +35,6 @@ inline int64_t start_index(int64_t a, int64_t b, int64_t c) {
 
 inline int64_t end_index(int64_t a, int64_t b, int64_t c) {
   return 1 + ((a + 1) * c - 1) / b;
-}
-
-inline void adaptive_pool_empty_output_check(const Tensor& gradOutput_, const char* arg_name) {
-  int64_t ndim = gradOutput_.ndimension();
-  for (const auto i : c10::irange(1, ndim)) {
-    TORCH_CHECK(gradOutput_.size(i) > 0,
-      arg_name, "(): Expected grad_output to have non-zero size for non-batch dimensions, "
-      "but grad_output has sizes ", gradOutput_.sizes(), " with dimension ", i,
-      " being empty");
-  }
 }
 
 } // namespace at::native
