+++ aten/src/ATen/native/FractionalMaxPooling.h
@@ -1,6 +1,7 @@
 #pragma once
 #include <ATen/core/Tensor.h>
 #include <ATen/TensorUtils.h>
+#include <ATen/native/PoolingChecks.h>
 #include <c10/util/irange.h>
 
 namespace at::native {
@@ -25,56 +26,6 @@ inline std::vector<int64_t> generate_intervals(
     sequence[outputSize - 1] = inputSize - poolSize;
   }
   return sequence;
-}
-
-template <int64_t ndim>
-inline void fractional_max_pool_check_shape(
-    const Tensor& input,
-    const Tensor& randomSamples) {
-
-  TORCH_CHECK(
-      input.scalar_type() == randomSamples.scalar_type(),
-      "Expect _random_samples to have the same dtype as input");
-
-  int64_t ndimension = randomSamples.ndimension();
-  TORCH_CHECK(
-      ndimension == 3,
-      "Expect _random_samples to have 3 dimensions, got ", ndimension);
-
-  int64_t N = randomSamples.size(0);
-  int64_t C = randomSamples.size(1);
-  int64_t D = randomSamples.size(2);
-
-  int64_t input_batch = 0, input_channel = 0;
-  if (ndim == 2) {
-    // fractional_max_pool2d
-    if (input.ndimension() == 3) {
-      input_batch = 1;
-      input_channel = input.size(0);
-    } else {
-      input_batch = input.size(0);
-      input_channel = input.size(1);
-    }
-  } else {
-    // factional_max_pool3d
-    if (input.ndimension() == 4) {
-      input_batch = 1;
-      input_channel = input.size(0);
-    } else {
-      input_batch = input.size(0);
-      input_channel = input.size(1);
-    }
-  }
-
-  TORCH_CHECK(
-      N >= input_batch,
-      "Expect _random_samples.size(0) no less then input batch size.");
-  TORCH_CHECK(
-      C == input_channel,
-      "Expect _random_samples.size(1) equals to input channel size.");
-  TORCH_CHECK(
-      D == ndim,
-      "Expect _random_samples.size(2) equals to ", ndim, "; got ", D, ".");
 }
 
 } // namespace at::native
